+++ src/test/java/com/airdropmc/config/DropOptionsTest.java
@@ -184,7 +184,7 @@ class DropOptionsTest {
     void getDropHeight_rejectsNegativeExplicitValue() {
         DropOptions options = DropOptions.createDefault().withDropHeight(-5);
 
-        assertEquals(40, options.getDropHeight());
+        assertEquals(100, options.getDropHeight());
     }
 
     @Test